fix(charts/paperless-ngx): unused Redis `existingSecret` as in #8,#9,#10
The behavior of the `existingSecret` for Redis was confusing before: Only the password from the secret was used and our own `Secret` containing the URI was (supposedly) still rendered.  This was discovered to be false it wasn't rendered if `existingSecret` was set.

Now the `existingSecret` defines the name of a Secret containing a `uri` key with the fully qualified URI for your Redis instance, just like our own secret. This done to avoid Secret duplication within the cluster since ours isn't created if the property is supplied. Re-using values from existing Secrets to render new ones seems illogical.

- update `.tpl` named templates to remove unsued ones
- add `default` to use the provided value if set
- update documentation
- add default DB hosts etc. to `values` since subcharts are also enabled by default
- update `README` and the `values` schema