feat(charts/linkwarden)!: update `existingSecret` API
Removed the sub-properties `name` and `key` from the `existingSecret` object we're only ever interested in the database URI and providing a name entirely skips the creation of our Secret, regardless of configured key. This change is in line with commit 39cc55 which also introduces this configuration for the `paperless-ngx` chart.

Beware that this is a BREAKING change.