fix(charts/paperless-ngx): generate hostnames for Redis and PostgreSQL #14
This commit fixes the issue #14 by generating the hostnames for Redis and PostgreSQL based on the known suffixes and the user's release name ahead of time and injecting the new values into the Redis URI and Paperless' ConfigMap. We also do away with the defaults in for the respective properties in `values.yaml`  and new doc comments have been added to notify users that the key doesn't need a value.

- chart version bumped to `0.2.3`
- updated Artifacthub `changes` annotation
- regenerated docs and `schema.json`