chore: reformat scripts and docs
This reformats the scripts in line with the '.editorconfig' removal. We're foregoing the version bump for the 'keycloak-operator' chart since this only modifies a script used to retrieve CRDs which we ship with the chart, not the CRDs itself - therefore preserving immutability. Regardless I want to avoid unnecessary version bumps.